+++ tests/Function/Trait/SqlWalkerTestTrait.php
@@ -1,0 +1,44 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * Copyright (c) Precision Soft
+ */
+
+namespace PrecisionSoft\Doctrine\Utility\Test\Function\Trait;
+
+use Doctrine\DBAL\Connection;
+use Doctrine\DBAL\Platforms\MySQLPlatform;
+use Doctrine\DBAL\Platforms\PostgreSQLPlatform;
+use Doctrine\ORM\Query\SqlWalker;
+use Mockery;
+
+trait SqlWalkerTestTrait
+{
+    private function createMysqlSqlWalker(): SqlWalker|Mockery\MockInterface
+    {
+        $connection = Mockery::mock(Connection::class);
+        $connection->shouldReceive('getDatabasePlatform')
+            ->andReturn(new MySQLPlatform());
+
+        $sqlWalker = Mockery::mock(SqlWalker::class);
+        $sqlWalker->shouldReceive('getConnection')
+            ->andReturn($connection);
+
+        return $sqlWalker;
+    }
+
+    private function createNonMysqlSqlWalker(): SqlWalker|Mockery\MockInterface
+    {
+        $connection = Mockery::mock(Connection::class);
+        $connection->shouldReceive('getDatabasePlatform')
+            ->andReturn(new PostgreSQLPlatform());
+
+        $sqlWalker = Mockery::mock(SqlWalker::class);
+        $sqlWalker->shouldReceive('getConnection')
+            ->andReturn($connection);
+
+        return $sqlWalker;
+    }
+}